+++ src/main/java/com/example/springboot/service/OpportunityService.java
@@ -26,6 +26,10 @@ public class OpportunityService {
         return opportunityRepository.findAllWithStatus();
     }
 
+    public List<Map<String, Object>> getAnalysis() {
+        return opportunityRepository.getAnalysis();
+    }
+
     public List<Map<String, Object>> getAllDataOpportunities() {
         return opportunityRepository.findAllDataWithStatus();
     }
@@ -48,6 +52,7 @@ public class OpportunityService {
             existingOpportunity.setExpected_closing(opportunity.getExpected_closing());
             existingOpportunity.setTags(opportunity.getTags());
             existingOpportunity.setStage_id(opportunity.getStage_id());
+            existingOpportunity.setCreated_date(existingOpportunity.getCreated_date());
             existingOpportunity.setRating(opportunity.getRating());
             existingOpportunity.setStatus(opportunity.getStatus());
             return opportunityRepository.save(existingOpportunity);
@@ -59,9 +64,18 @@ public class OpportunityService {
         Map<String, String> result = new HashMap<>();
 
         try {
-            opportunityRepository.deleteByIdWithStatus(id);
+            List<Map<String, Object>> check = opportunityRepository.check(id);
 
-            result.put("msg", "The opportunity is deleted successfully.");
+            if (check.size() == 0 || check == null) {
+                opportunityRepository.deleteByIdWithStatus(id);
+
+                result.put("flag", "0");
+                result.put("msg", "The opportunity is deleted successfully.");
+            } else {
+                result.put("flag", "1");
+                result.put("msg", "Cannot delete");
+            }
+
             return ResponseEntity.ok(result);
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Map.of("error", "An error occurred while deleting the opportunity"));
